Fix Linux CE.
GitOrigin-RevId: 7758d3fe4cc16210338f739e2dcc748d55e808a5

+++ tdutils/td/utils/port/detail/Epoll.cpp
@@ -112,7 +112,7 @@ void Epoll::run(int timeout_ms) {
       flags = flags | PollFlags::Error();
     }
     if (event->events) {
-      LOG(FATAL) << "Unsupported epoll events: " << event->events;
+      LOG(FATAL) << "Unsupported epoll events: " << static_cast<int32>(event->events);
     }
     //LOG(DEBUG) << "Epoll event " << tag("fd", event->data.fd) << tag("flags", format::as_binary(flags));
     auto pollable_fd = PollableFd::from_list_node(static_cast<ListNode *>(event->data.ptr));
